Overhauling unit tests to use local fake service.
We don't want to connect to the live IEX Cloud service for unit tests,
especially when the test itself requires a paid subscription to run.
Using local fakes avoids external dependencies and ensures anyone can
run the tests.

This is the first of several changes to rewrite existing tests to use
the new fake.